langchain-to-deepagents-in-action/code-examples/part1-langchain-basics/requirements.txt
# TechAssist v0.1-v0.3 依賴
langchain>=0.2.0
langchain-anthropic>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.25.0
